+++ R/BiCopGradMMD.R
@@ -40,7 +40,7 @@
 #'     \item \code{inv.l1}: \eqn{k(x,y) = 1/(1+\|\frac{x-y}{\gamma}\|_1)^\alpha
 #'     }{k(x,y) = 1 / ( 1 + || (x-y) / gamma ||_1 )^\alpha}
 #'   }
-#'  Each of these names can receive the suffix ".KG", such as "gaussian.KG"
+#'  Each of these names can receive the suffix ".Phi", such as "gaussian.Phi"
 #'  to indicates that the kernel \eqn{k(x,y)} is replaced by
 #'  \eqn{k(\Phi^{-1}(x) , \Phi^{-1}(y))} where \eqn{\Phi^{-1}} denotes the quantile
 #'  function of the standard Normal distribution.
@@ -104,7 +104,7 @@
 #'
 BiCopGradMMD <- function(
   u1, u2, family, tau, par = NULL, par2=0,
-  kernel = "gaussian.KG", gamma=0.95, alpha=1, epsilon=0.0001,
+  kernel = "gaussian.Phi", gamma=0.95, alpha=1, epsilon=0.0001,
   method = "QMCV", quasiRNG = "sobol", ndrawings=10)
 {
   if (!is.null(par))