+++ NEWS.md
@@ -1,3 +1,13 @@
+
+# MMDCopula 0.1.1
+
+* Change email address of Alexis Derumigny.
+* Vectorized version of `BiCopSim.MO`.
+* `BiCopSim.MO` now returns a list with the estimated parameter and Kendall's tau for consistency with the other `BiCopEstMMD` function.
+* New function `BiCopMMDConfInt` for bootstrapped- and subsampling-based confidence intervals.
+* Renaming of kernels such as `gaussian.KG` to `gaussian.Phi` for coherence.
+
+
 # MMDCopula 0.1.0
 
 * Initial release